pyproject.toml
[project]
name = "webscraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
    "selenium>=4.41.0",
    "setuptools>=82.0.1",
    "spotipy>=2.24.0",
    "undetected-chromedriver>=3.5.5",
    "webdriver-manager>=4.0.2",
]
